[REF] Initial refactoring push.
