--- brainstat/stats/__init__.py
@@ -1,16 +1,0 @@
-from .SurfStatAvSurf import SurfStatAvSurf
-from .SurfStatAvVol import SurfStatAvVol
-from .SurfStatEdg import SurfStatEdg
-from .SurfStatF import SurfStatF
-from .SurfStatLinMod import SurfStatLinMod
-from .SurfStatNorm import SurfStatNorm
-from .SurfStatP import SurfStatP
-from .SurfStatPeakClus import SurfStatPeakClus
-from .SurfStatQ import SurfStatQ
-from .SurfStatResels import SurfStatResels
-from .SurfStatSmooth import SurfStatSmooth
-from .SurfStatStand import SurfStatStand
-from .stat_threshold import stat_threshold
-from .SurfStatT import SurfStatT
-from .term import Term
-from .term import Random
